Cast atol to u_int to avoid this warning:
test/checkpoint/workers.c:152:27: error: implicit conversion loses integer precision: 'long' to 'u_int' (aka 'unsigned int') [-Werror,-Wshorten-64-to-32]

+++ test/checkpoint/checkpointer.c
@@ -231,7 +231,7 @@ get_key_int(WT_CURSOR *cursor, int table_index, u_int *rval)
 		cursor->get_key(cursor, &key);
 		memset(buf, 0, 128);
 		memcpy(buf, key.data, key.size);
-		val = atol(buf);
+		val = (u_int)atol(buf);
 	}
 
 	*rval = val;
